# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Endpoints for Go Server",
        "title": "Go Rest Api",
        "contact": {}
    },
    "paths": {
        "/api/book/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Add a specific book",
                "parameters": [
                    {
                        "description": "Book title",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book author",
                        "name": "author",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book coverUrl",
                        "name": "coverUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book post url",
                        "name": "postUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IDResponse"
                        }
                    }
                }
            }
        },
        "/api/book/delete/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Delete a specific book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetBookResponse"
                        }
                    }
                }
            }
        },
        "/api/book/update": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Update a specific book",
                "parameters": [
                    {
                        "description": "Book title",
                        "name": "title",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book author",
                        "name": "author",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book coverUrl",
                        "name": "coverUrl",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Book post url",
                        "name": "postUrl",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IDResponse"
                        }
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Get a specific book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetBookResponse"
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.GetBookResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.GetBookResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "coverUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "postUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}